refactor evm logs
