+++ x/cronos/keeper/evmhandlers/send_to_ibc.go
@@ -1,0 +1,113 @@
+package evmhandler
+
+import (
+	"fmt"
+	"math/big"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+
+	cronoskeeper "github.com/crypto-org-chain/cronos/x/cronos/keeper"
+	"github.com/crypto-org-chain/cronos/x/cronos/types"
+)
+
+var _ types.EvmLogHandler = SendToIbcHandler{}
+
+const SendToIbcEventName = "__CronosSendToIbc"
+
+// SendToIbcEvent represent the signature of
+// `event __CronosSendToIbc(string recipient, uint256 amount)`
+var SendToIbcEvent abi.Event
+
+func init() {
+	addressType, _ := abi.NewType("address", "", nil)
+	uint256Type, _ := abi.NewType("uint256", "", nil)
+	stringType, _ := abi.NewType("string", "", nil)
+
+	SendToIbcEvent = abi.NewEvent(
+		SendToIbcEventName,
+		SendToIbcEventName,
+		false,
+		abi.Arguments{abi.Argument{
+			Name:    "sender",
+			Type:    addressType,
+			Indexed: false,
+		}, abi.Argument{
+			Name:    "recipient",
+			Type:    stringType,
+			Indexed: false,
+		}, abi.Argument{
+			Name:    "amount",
+			Type:    uint256Type,
+			Indexed: false,
+		}},
+	)
+}
+
+// SendToIbcHandler handles `__CronosSendToIbc` log
+type SendToIbcHandler struct {
+	bankKeeper   types.BankKeeper
+	cronosKeeper cronoskeeper.Keeper
+}
+
+func NewSendToIbcHandler(bankKeeper types.BankKeeper, cronosKeeper cronoskeeper.Keeper) *SendToIbcHandler {
+	return &SendToIbcHandler{
+		bankKeeper:   bankKeeper,
+		cronosKeeper: cronosKeeper,
+	}
+}
+
+func (h SendToIbcHandler) EventID() common.Hash {
+	return SendToIbcEvent.ID
+}
+
+func (h SendToIbcHandler) Handle(
+	ctx sdk.Context,
+	contract common.Address,
+	data []byte,
+	_ func(contractAddress common.Address, logSig common.Hash, logData []byte),
+) error {
+	unpacked, err := SendToIbcEvent.Inputs.Unpack(data)
+	if err != nil {
+		// log and ignore
+		h.cronosKeeper.Logger(ctx).Info("log signature matches but failed to decode")
+		return nil
+	}
+
+	denom, found := h.cronosKeeper.GetDenomByContract(ctx, contract)
+	if !found {
+		return fmt.Errorf("contract %s is not connected to native token", contract)
+	}
+
+	if !types.IsValidIBCDenom(denom) && !types.IsValidCronosDenom(denom) {
+		return fmt.Errorf("the native token associated with the contract %s is neither an ibc voucher or a cronos token", contract)
+	}
+
+	contractAddr := sdk.AccAddress(contract.Bytes())
+	sender := sdk.AccAddress(unpacked[0].(common.Address).Bytes())
+	recipient := unpacked[1].(string)
+	amount := sdk.NewIntFromBigInt(unpacked[2].(*big.Int))
+	coins := sdk.NewCoins(sdk.NewCoin(denom, amount))
+
+	if types.IsSourceCoin(denom) {
+		// it is a source token, we need to mint coins
+		if err = h.bankKeeper.MintCoins(ctx, types.ModuleName, coins); err != nil {
+			return err
+		}
+		// send the coin to the user
+		if err = h.bankKeeper.SendCoinsFromModuleToAccount(ctx, types.ModuleName, sender, coins); err != nil {
+			return err
+		}
+	} else {
+		// First, transfer IBC coin to user so that he will be the refunded address if transfer fails
+		if err = h.bankKeeper.SendCoins(ctx, contractAddr, sender, coins); err != nil {
+			return err
+		}
+	}
+	// Initiate IBC transfer from sender account
+	if err = h.cronosKeeper.IbcTransferCoins(ctx, sender.String(), recipient, coins); err != nil {
+		return err
+	}
+	return nil
+}
